requirements.txt
-i https://pypi.org/simple
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
isort==6.0.1; python_full_version >= '3.9.0'
lxml==5.3.1; python_version >= '3.6'
pygame-ce==2.5.3; python_version >= '3.9'
requests==2.32.3; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'
